Moved files and updated namespaces
