+++ LaunchboxPluginsTests/OnlineVideoLinks/GameVideoTests.cs
@@ -20,8 +20,8 @@
 
 using LaunchboxPluginsTests.MockedClasses;
 using NSubstitute;
-using OnlineVideoLinks;
 using OnlineVideoLinks.Models;
+using OnlineVideoLinks.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
