+++ OnlineVideoLinks/Models/GameVideo.cs
@@ -18,6 +18,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
+using OnlineVideoLinks.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
