+++ OnlineVideoLinks/Utilities/VlcUtilities.cs
@@ -27,7 +27,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace OnlineVideoLinks
+namespace OnlineVideoLinks.Utilities
 {
     public class VlcUtilities
     {
